SnapHook/src/requirements.txt
fastapi
uvicorn[standard]
httpx
PyYAML
